component reorganization, start of making responsive
