+++ src/App.css
@@ -8,12 +8,10 @@
 }
 
 .container {
-  margin-left: 5%;
-  margin-right: 5%;
-  margin-top: 7%;
-  width: 260px !important;
-  margin-left: auto;
-  margin-right: auto;
+  margin-top: 5vh;
+  overflow-x: hidden;
+  display: flex;
+  flex-direction: column;
 }
 
 .center-wrapper {
@@ -25,8 +23,11 @@
   width: 122px;
   height: 0px;
   border: 0.7px solid #000000;
+}
 
-  margin: 30% 0 30% 0;
+.horizontal-scroll{
+  overflow-x: scroll;
+  display: flex;
 }
 
 .pf-c-button{
@@ -50,10 +51,4 @@
   --pf-global--BorderColor--300 : green;
   --pf-global--BorderColor-100: green;
   --pf-global--BorderColor--dark-100: green;
-}
-
-.horizontal-scroll{
-  overflow-x:scroll ;
-  width:260px;
-  display: flex;
 }
