+++ src/screens/AddressVerify.tsx
@@ -1,10 +1,10 @@
 import * as React from "react";
-import AddressCheckBox from "../components/AddressCheckBox";
-import AddressCheckBoxLoading from "../components/AddressCheckBoxLoading";
-import AddressErrorBox from "../components/AddressErrorBox";
-import AddressInvalidBox from "../components/AddressInvalidBox";
-import AddressAPIErrorBox from "../components/AddressAPIErrorBox";
-import StateSelection from "../components/StateSelection";
+import AddressCheckBox from "../components/address/AddressCheckBox";
+import AddressCheckBoxLoading from "../components/address/AddressCheckBoxLoading";
+import AddressErrorBox from "../components/address/AddressErrorBox";
+import AddressInvalidBox from "../components/address/AddressInvalidBox";
+import AddressAPIErrorBox from "../components/address/AddressAPIErrorBox";
+import StateSelection from "../components/address/StateSelection";
 import { TextInput, Button } from "@patternfly/react-core";
 import { useNavigate } from "react-router-dom";
